requirements.txt
discord.py>=2.6.0
asyncpg>=0.27.0
python-dotenv>=1.0.0
